# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task Auth API",
    "version": "1.0.0",
    "description": "Email authentication and protected task-service endpoints using Supabase Auth."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "tags": [
    { "name": "Auth", "description": "Sign up, log in, and log out" },
    { "name": "Public", "description": "Endpoints that do not require authentication" },
    { "name": "Protected", "description": "Endpoints that require a valid Supabase access token" }
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Create an account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Credentials" },
              "example": { "email": "person@example.com", "password": "your-password" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Supabase user created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log in with email and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Credentials" },
              "example": { "email": "person@example.com", "password": "your-password" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Supabase session tokens",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenPair" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log out the current session",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Session signed out" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/public/info": {
      "get": {
        "tags": ["Public"],
        "summary": "Get public information",
        "responses": {
          "200": {
            "description": "Public welcome message",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PublicInfo" },
                "example": { "message": "Welcome stranger! This info is public." }
              }
            }
          }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "tags": ["Protected"],
        "summary": "Get the authenticated user's safe profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Verified user metadata",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Profile" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "tags": ["Protected"],
        "summary": "Get the authenticated user's dashboard",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard welcome response" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the access_token returned by POST /auth/login."
      }
    },
    "schemas": {
      "Credentials": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "format": "password" }
        }
      },
      "TokenPair": {
        "type": "object",
        "required": ["access_token", "refresh_token"],
        "properties": {
          "access_token": { "type": "string" },
          "refresh_token": { "type": "string" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Profile": {
        "type": "object",
        "required": ["id", "created_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "PublicInfo": {
        "type": "object",
        "required": ["message"],
        "properties": { "message": { "type": "string" } }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "type": "string" } }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Missing or invalid request fields",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "Missing, invalid, or expired access token",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  }
}